# spring-server-generated/src/main/java/io/swagger/api/PromotionApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.13).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package io.swagger.api;

import io.swagger.model.Promotion;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2019-11-12T07:56:28.663Z[GMT]")
@Api(value = "promotion", description = "the promotion API")
public interface PromotionApi {

  @ApiOperation(value = "add a valid promotion code with corresponding discount in database.", nickname = "addPromotion", notes = "add a new promotion code with discount.", tags={ "admins", })
  @ApiResponses(value = {
      @ApiResponse(code = 200, message = "discount for promotion code created. or an existing item already exist, override it"),
      @ApiResponse(code = 400, message = "invalid input, object invalid"),})
  @RequestMapping(value = "/promotion",
      consumes = { "application/json" },
      method = RequestMethod.POST)
  ResponseEntity<Void> addPromotion(@ApiParam(value = "promotion code item to add"  )  @Valid @RequestBody Promotion body);


  @ApiOperation(value = "deletes a promotion code", nickname = "deletePromotion", notes = "Delete an item in the system", tags={ "admins", })
  @ApiResponses(value = {
      @ApiResponse(code = 200, message = "item deleted"),
      @ApiResponse(code = 400, message = "invalid input, object invalid"),
      @ApiResponse(code = 404, message = "an existing item doesn't exist") })
  @RequestMapping(value = "/promotion",
      method = RequestMethod.DELETE)
  ResponseEntity<Void> deletePromotion(@NotNull @ApiParam(value = "pass a promotion code string for looking up a promotion", required = true) @Valid @RequestParam(value = "searchCode", required = true) String searchCode);


  @ApiOperation(value = "Take in an promotion code and return the discount.", nickname = "getPromotion", notes = "Get discount by promotion code ", response = Promotion.class, tags={ "developers", })
  @ApiResponses(value = {
      @ApiResponse(code = 200, message = "discount after discount is succesfully returned.", response = Promotion.class),
      @ApiResponse(code = 400, message = "invalid input, object invalid"),
      @ApiResponse(code = 404, message = "invalid input, object invalid")})
  @RequestMapping(value = "/promotion",
      produces = { "application/json" },
      method = RequestMethod.GET)
  ResponseEntity<Promotion> getPromotion(@NotNull @ApiParam(value = "promotion code to use.", required = true) @Valid @RequestParam(value = "code", required = true) String code);

}