# spring-server-generated/src/main/resources/api/api.yaml
swagger: '2.0'
info:
  version: '1.0'
  title: 'Pizza-Pizza'
  description: 'This is the API doc for Pizza-Pizza store.'
  contact:
    email: fan.fan@husky.neu.edu

  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html

# tags are used for organizing operations
...
paths:
...
  /promotion:
    get:
      parameters:
        - in: query
          name: code
          required: true
          schema:
            type: string
          description: promotion code to use.
      tags:
        - developers
      summary: Take in an promotion code and return the discount.
      operationId: getPromotion
      description: |
        Get discount by promotion code
      produces:
        - application/json
      responses:
        200:
          description: discount after discount is succesfully returned.
          schema:
            $ref: '#/definitions/Promotion'
        400:
          description: invalid input, object invalid

    post:
      tags:
        - admins
      summary: add a valid promotion code with corresponding discount in database.
      operationId: addPromotion
      description: add a new promotion code with discount.
      responses:
        201:
          description: discount for promotion code created.
        400:
          description: invalid input, object invalid
        409:
          description: an existing item doesn't exist
      parameters:
        - in: body
          name: promotion
          description: promotion code item to add
          schema:
            $ref: '#/definitions/Promotion'

    delete:
      tags:
        - admins
      summary: deletes a promotion code
      operationId: deletePromotion
      description: Delete an item in the system
      responses:
        201:
          description: item deleted
        400:
          description: 'invalid input, object invalid'
        409:
          description: an existing item doesn't exist
      parameters:
        - in: query
          name: searchCode
          description: pass a promotion code string for looking up a promotion
          required: true
          schema:
            type: string
definitions:
...
  Promotion:
    type: object
    properties:
      code:
        type: string
      discount:
        type: number
# Added by API Auto Mocking Plugin
host: virtserver.swaggerhub.com
basePath: /ffhere/Pizza-Pizza/1.0
schemes:
  - https